# platform/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI4S M0 entry and scope
  version: 0.2.0
  description: Same-origin session and CSRF. All project mutations are actor-scoped idempotent revision commands. Later-stage submissions are exploratory only.
security: [{session: []}]
paths:
...
  /api/v1/projects/{project_id}/status:
    parameters: [{$ref: '#/components/parameters/Project'}]
    get:
      responses: {'200': {description: "Project metadata and nonauthoritative workflow"}, '404': {$ref: '#/components/responses/Error'}}
  /api/v1/projects/{project_id}/scope:
    parameters: [{$ref: '#/components/parameters/Project'}]
    get:
      responses:
        '200': {description: "Current scope, version history, revision, fixed groups and provider availability", content: {application/json: {schema: {$ref: '#/components/schemas/ScopeResponse'}}}}
        '404': {$ref: '#/components/responses/Error'}
    post:
      description: Owner/researcher only. Confirm freezes validated content; edit clones a confirmed version. Existing candidate provenance and conflict identities cannot be removed.
      parameters: [{$ref: '#/components/parameters/Key'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - {$ref: '#/components/schemas/Command'}
                - type: object
                  required: [action]
                  properties:
                    action: {enum: [save, confirm, edit]}
                    direction: {type: string, minLength: 1, maxLength: 2000}
                    details: {$ref: '#/components/schemas/ScopeDetails'}
      responses:
        '200': {description: "Saved scope and history", content: {application/json: {schema: {$ref: '#/components/schemas/ScopeResponse'}}}}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
...
  /api/v1/projects/{project_id}/evidence:
    parameters: [{$ref: '#/components/parameters/Project'}]
    post:
      parameters: [{$ref: '#/components/parameters/Key'}]
      requestBody: {$ref: '#/components/requestBodies/Command'}
      responses: {'200': {description: "Exploratory location assertion; submitted human_verified cannot create authoritative verification"}, '422': {$ref: '#/components/responses/Error'}}
...
components:
  securitySchemes:
    session: {type: apiKey, in: cookie, name: sessionid}
  parameters:
    Project: {name: project_id, in: path, required: true, schema: {type: string, format: uuid}}
...
    Key: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 160}}
  responses:
    Error:
      description: Structured error; private missing/denied projects share 404. Anonymous is 403 NOT_AUTHENTICATED; CSRF_FAILED is distinct.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message, owner, recovery, details, trace_id]
                properties: {code: {type: string}, message: {type: string}, owner: {type: string}, recovery: {type: string}, details: {type: object}, trace_id: {type: string}}
  requestBodies:
    Command:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/Command'}}}
  schemas:
    Command:
      type: object
      required: [expected_revision]
      properties: {expected_revision: {type: integer, minimum: 0}}
    ScopeDetails:
      type: object
      required: [answers, candidates, conflicts, semantic_review]
      properties:
        answers:
          type: object
          additionalProperties: false
          properties:
            object: {type: string, maxLength: 2000}
            mechanism: {type: string, maxLength: 2000}
            method: {type: string, maxLength: 2000}
            outcome: {type: string, maxLength: 2000}
            context: {type: string, maxLength: 2000}
            years: {type: string, description: Start/end years between 1800 and current year plus one}
            languages: {type: string, maxLength: 2000}
            types: {type: string, maxLength: 2000}
            include: {type: string, maxLength: 2000}
            exclude: {type: string, maxLength: 2000}
            resources: {type: string, maxLength: 2000}
        candidates:
          type: array
          maxItems: 200
          items:
            type: object
            required: [id, term, source, decision]
            properties:
              id: {type: string, maxLength: 80}
              term: {type: string, maxLength: 200}
              source: {enum: [manual, project_keyword], description: Only server may seed project_keyword}
              decision: {enum: [pending, accepted, rejected, replaced]}
              replacement: {type: string, maxLength: 200}
        conflicts:
          type: array
          maxItems: 200
          items:
            type: object
            required: [id, description, resolution]
            properties: {id: {type: string, maxLength: 80}, description: {type: string, maxLength: 2000}, resolution: {type: string, maxLength: 2000}}
        semantic_review: {type: string, maxLength: 4000, description: Required on confirm; manual statement not AI certainty}
    Scope:
      type: object
      properties:
        id: {type: string, format: uuid}
        version: {type: integer, minimum: 1}
        status: {enum: [draft, confirmed]}
        direction: {type: string}
        core_keywords: {type: array, items: {type: string}}
        details: {$ref: '#/components/schemas/ScopeDetails'}
        confirmed_at: {type: [string, 'null'], format: date-time}
    ScopeResponse:
      type: object
      properties:
        project_id: {type: string, format: uuid}
        project_revision: {type: integer}
        scope: {$ref: '#/components/schemas/Scope'}
        history: {type: array, items: {$ref: '#/components/schemas/Scope'}}
        groups: {type: array, minItems: 4, maxItems: 4, items: {type: object}}
        providers: {type: object, properties: {ai: {const: UNCONFIGURED}, knowledge_graph: {const: UNCONFIGURED}}}